task/requirements.txt
gunicorn==23.0.0
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
factory_boy==3.3.1
